requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
requests
feedparser
nltk
apscheduler
tzdata
loguru
streamlit
plotly
python-dotenv